fix: split script execution helpers
